fix(ci): fait passer la suite des epics 4 à 11 sous MySQL et compte restreint
Le job « PHPUnit (MySQL 8) » n'avait jamais été vert depuis l'arrivée des epics 4
à 11 : 469 tests en échec sur 974. Ces tests n'avaient été éprouvés que sur SQLite
en mémoire, où la suite est verte, et le job MySQL exerce une contrainte que SQLite
ignore — la séparation des comptes de `docs/ops/database-users.md`. Les échecs ne
venaient donc pas du code applicatif mais de deux angles morts.

**Le DDL passait par le compte applicatif (368 échecs).** Quarante-neuf fichiers
employaient `RefreshDatabase`, qui lance `migrate:fresh` sur la connexion par
défaut — celle du compte restreint, qui n'a ni `DROP` ni `CREATE`. Le refus
`DROP command denied` n'était pas un défaut de configuration mais la preuve que la
matrice tient. Le nouveau trait `RefreshesSeparatedDatabase` confie le DDL à la
connexion de migration et laisse les tests s'exécuter sous le compte restreint,
reproduisant l'asymétrie de production. Sur SQLite en mémoire, où il n'y a pas deux
comptes à séparer, le comportement d'origine est conservé.

**Quarante-trois tables métier n'avaient pas de `GRANT UPDATE` (152 échecs).** La
règle du projet veut que la migration qui crée une table l'accorde elle-même :
`SELECT` et `INSERT` sont hérités du schéma, `UPDATE` jamais. Les dix-sept
migrations des epics 2 et 3 le font ; les quatorze d'août 2026 l'ont oublié, si
bien que toute écriture — et jusqu'au moindre `SELECT ... FOR UPDATE`, qui exige ce
privilège — était refusée. Ces migrations ayant déjà tourné en production, le
rattrapage passe par une migration idempotente plutôt que par leur réécriture.
`account_movements`, les tables de versions et d'historiques en restent absentes,
comme `audit_logs` et `user_history` : leur immuabilité est le sujet, pas un
obstacle. Aucun `DELETE` n'est accordé. La matrice écrite est mise à jour pour la
production et la préproduction, avec la marche à suivre pour l'exploitant.

**Les preuves de concurrence ne se nettoyaient pas.** Elles committent leurs
fixtures pour que deux processus se voient réellement, mais nettoyaient par
`DELETE` sous le compte applicatif, qui n'en a pas le droit ; leur résidu faussait
ensuite une quinzaine de comptages ailleurs dans la suite. Le ménage passe
désormais par la connexion de migration, et comme les traces d'audit sont
indélébiles — aucun `DELETE` ne les atteint, et il n'était pas question
d'affaiblir cette garantie — la remise à zéro est structurelle et réservée au
compte privilégié.

Deux tests restent rouges, tous deux antérieurs à ce commit et liés à la même
tension entre fixtures committées et isolation transactionnelle :

  • `UserHistoryTest::test_ac_4_mysql_triggers_refuse_privileged_update_and_delete`
    attend l'erreur 1644 du déclencheur et reçoit 1205, un délai de verrou : la
    ligne visée vit dans la transaction non committée du test, que le compte
    privilégié ne peut pas atteindre.
  • `ExpenseApprovalConcurrencyDatabaseTest` voit ses deux processus fils échouer
    techniquement.

Mesuré sur un MySQL 8.0.44 provisionné comme le job CI, comptes et privilèges
identiques : 970 tests passent contre 534 au départ. SQLite reste vert sans
régression (943 passés, 31 ignorés), Pint, les tests JS, le build et le budget de
bundle également.

+++ tests/Feature/ShareEntitlementLifecycleTest.php
@@ -45,9 +45,9 @@ class ShareEntitlementLifecycleTest extends IdentityTestCase
 
         $this->assertSame([50_000, 300_000, 150_000], $firstAmounts);
         $this->assertSame($firstAmounts, $secondAmounts);
-        $this->assertSame(1_000_000, ShareEntitlement::query()->sum('share_amount'));
-        $this->assertSame(100_000, ShareEntitlement::query()->where('beneficiary_id', $contributor->getKey())->sum('share_amount'));
-        $this->assertSame(300_000, ShareEntitlement::query()->where('beneficiary_id', $executor->getKey())->sum('share_amount'));
+        $this->assertSame(1_000_000, (int) ShareEntitlement::query()->sum('share_amount'));
+        $this->assertSame(100_000, (int) ShareEntitlement::query()->where('beneficiary_id', $contributor->getKey())->sum('share_amount'));
+        $this->assertSame(300_000, (int) ShareEntitlement::query()->where('beneficiary_id', $executor->getKey())->sum('share_amount'));
     }
 
     public function test_ac_46_abandoned_half_paid_contract_has_only_half_of_all_rights(): void
@@ -55,7 +55,7 @@ class ShareEntitlementLifecycleTest extends IdentityTestCase
         [$actor, $client, $contract, $invoice, $account] = $this->context();
         app(PaymentService::class)->record($this->payload($client, $contract, $invoice, $account, 500_000), $actor);
 
-        $this->assertSame(500_000, ShareEntitlement::query()->sum('share_amount'));
+        $this->assertSame(500_000, (int) ShareEntitlement::query()->sum('share_amount'));
         $this->assertSame(500_000, $invoice->refresh()->outstandingAmount());
     }
